requirements.txt
glfw==2.5.3
numba==0.58.1
numpy==1.26.2
opensimplex==0.4.4
Pillow==9.1.1
Pillow==10.1.0
pygame==2.1.2
PyGLM==2.5.7
PyOpenGL==3.1.6
PyOpenGL==3.1.7
